refactor: move pre-send info to provider
